requirements.txt
python-dotenv
ollama
requests
ipython
beautifulsoup4
openai
gradio